+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## UNRELEASED
+
+- Implement CLI workflow. User will use script `spectacle-cli init` to kick off the workflow to create a server or build directory.
+
 ## 0.6.4
 
 - Update to `spectacle@8`.